transcribe-long/pyproject.toml
[project]
name = "transcribe-long"
version = "0.1.0"
description = "Transcribe large audio files by splitting on silence, transcribing chunks, and stitching"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.81.0",
    "thds-core>=1.50.20260117040119",
    "thds-mops>=3.13.20260117040131",
]

[project.scripts]
transcribe = "transcribe.__main__:main"
transcribe-diarize = "transcribe.diarize.__main__:main"

[project.optional-dependencies]
diarize = [
    "pyannote-audio>=3.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 108

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
exclude = [
    '^venv/',
    '^tests/'
]
